ai-relay-server/requirements.txt
fastapi
uvicorn[standard]
httpx
websockets
deepgram-sdk
pyaudio
playsound3==2.2.1
sounddevice==0.4.7
numpy==2.0.1